+++ src/rover.c
@@ -1,5 +1,5 @@
-
 #include "config.h"
+#include "ui_funcs.h"
 
 void init_marks(Marks *marks)
 {
@@ -119,9 +119,6 @@ void disable_handlers()
 	sigaction(SIGWINCH, &sa, NULL);
 }
 
-void reload();
-void update_view();
-
 /* Handle any signals received since last call. */
 void sync_signals()
 {
@@ -164,22 +161,6 @@ int rover_get_wch(wint_t *wch)
 	while ((ret = get_wch(wch)) == (wint_t)ERR)
 		sync_signals();
 	return ret;
-}
-
-/* Get user programs from the environment. */
-
-#define ROVER_ENV(dst, src)                    \
-	if ((dst = getenv("ROVER_" #src)) == NULL) \
-		dst = getenv(#src);
-
-void get_user_programs()
-{
-	ROVER_ENV(user_shell, SHELL)
-	ROVER_ENV(user_pager, PAGER)
-	ROVER_ENV(user_editor, VISUAL)
-	if (!user_editor)
-		ROVER_ENV(user_editor, EDITOR)
-	ROVER_ENV(user_open, OPEN)
 }
 
 /* Do a fork-exec to external program (e.g. $EDITOR). */
@@ -229,12 +210,14 @@ done:
 
 int open_with_env(char *program, char *path)
 {
+	char buffer[PATH_MAX];
+
 	if (program) {
 #ifdef RV_SHELL
-		strncpy(BUF1, program, BUFLEN - 1);
-		strncat(BUF1, " ", BUFLEN - strlen(program) - 1);
-		shell_escaped_cat(BUF1, path, BUFLEN - strlen(program) - 2);
-		spawn((char *[]){ RV_SHELL, "-c", BUF1, NULL });
+		strncpy(buffer, program, PATH_MAX -1);
+		strncat(buffer, " ", PATH_MAX - strlen(program) - 1);
+		shell_escaped_cat(buffer, path, PATH_MAX - strlen(program) -2);
+		spawn((char *[]){ RV_SHELL, "-c", buffer, NULL });
 #else
 		spawn((char *[]){ program, path, NULL });
 #endif
@@ -243,47 +226,13 @@ int open_with_env(char *program, char *path)
 	return 0;
 }
 
-/* Curses setup. */
-void init_term()
-{
-	setlocale(LC_ALL, "");
-	initscr();
-	cbreak(); /* Get one character at a time. */
-	timeout(100); /* For getch(). */
-	noecho();
-	nonl(); /* No NL->CR/NL on output. */
-	intrflush(stdscr, FALSE);
-	keypad(stdscr, TRUE);
-	curs_set(FALSE); /* Hide blinking cursor. */
-	if (has_colors()) {
-		short bg;
-		start_color();
-#ifdef NCURSES_EXT_FUNCS
-		use_default_colors();
-		bg = -1;
-#else
-		bg = COLOR_BLACK;
-#endif
-		init_pair(RED, COLOR_RED, bg);
-		init_pair(GREEN, COLOR_GREEN, bg);
-		init_pair(YELLOW, COLOR_YELLOW, bg);
-		init_pair(BLUE, COLOR_BLUE, bg);
-		init_pair(CYAN, COLOR_CYAN, bg);
-		init_pair(MAGENTA, COLOR_MAGENTA, bg);
-		init_pair(WHITE, COLOR_WHITE, bg);
-		init_pair(BLACK, COLOR_BLACK, bg);
-	}
-	atexit((void (*)(void))endwin);
-	enable_handlers();
-}
 
 /* Update the listing view. */
 void update_view()
 {
-	int i, j;
-	int numsize;
-	int ishidden;
-	int marking;
+	int i, j, numsize, ishidden, marking;
+	char buffer_one[PATH_MAX], buffer_two[PATH_MAX];
+	wchar_t wbuffer[PATH_MAX];
 
 	mvhline(0, 0, ' ', COLS);
 	attr_on(A_BOLD, NULL);
@@ -291,14 +240,14 @@ void update_view()
 	mvaddch(0, COLS - 2, rover.tab + '0');
 	attr_off(A_BOLD, NULL);
 	if (rover.marks.nentries) {
-		numsize = snprintf(BUF1, BUFLEN, "%d", rover.marks.nentries);
+		numsize = snprintf(buffer_one, PATH_MAX, "%d", rover.marks.nentries);
 		color_set(RVC_MARKS, NULL);
-		mvaddstr(0, COLS - 3 - numsize, BUF1);
+		mvaddstr(0, COLS - 3 - numsize, buffer_one);
 	} else
 		numsize = -1;
 	color_set(RVC_CWD, NULL);
-	mbstowcs(WBUF, CWD, PATH_MAX);
-	mvaddnwstr(0, 0, WBUF, COLS - 4 - numsize);
+	mbstowcs(wbuffer, CWD, PATH_MAX);
+	mvaddnwstr(0, 0, wbuffer, COLS - 4 - numsize);
 	wcolor_set(rover.window, RVC_BORDER, NULL);
 	wborder(rover.window, 0, 0, 0, 0, 0, 0, 0, 0);
 	ESEL = MAX(MIN(ESEL, rover.nfiles - 1), 0);
@@ -334,27 +283,27 @@ void update_view()
 		else if (S_ISSOCK(EMODE(j)))
 			wcolor_set(rover.window, RVC_SOCK, NULL);
 		if (S_ISDIR(EMODE(j))) {
-			mbstowcs(WBUF, ENAME(j), PATH_MAX);
+			mbstowcs(wbuffer, ENAME(j), PATH_MAX);
 			if (ISLINK(j))
-				wcscat(WBUF, L"/");
+				wcscat(wbuffer, L"/");
 		} else {
 			char *suffix, *suffixes = "BKMGTPEZY";
 			off_t human_size = ESIZE(j) * 10;
-			int length       = mbstowcs(WBUF, ENAME(j), PATH_MAX);
-			int namecols     = wcswidth(WBUF, length);
+			int length       = mbstowcs(wbuffer, ENAME(j), PATH_MAX);
+			int namecols     = wcswidth(wbuffer, length);
 			for (suffix = suffixes; human_size >= 10240; suffix++)
 				human_size = (human_size + 512) / 1024;
 			if (*suffix == 'B')
-				swprintf(WBUF + length, PATH_MAX - length, L"%*d %c",
+				swprintf(wbuffer + length, PATH_MAX - length, L"%*d %c",
 				         (int)(COLS - namecols - 6),
 				         (int)human_size / 10, *suffix);
 			else
-				swprintf(WBUF + length, PATH_MAX - length, L"%*d.%d %c",
+				swprintf(wbuffer + length, PATH_MAX - length, L"%*d.%d %c",
 				         (int)(COLS - namecols - 8),
 				         (int)human_size / 10, (int)human_size % 10, *suffix);
 		}
 		mvwhline(rover.window, i + 1, 1, ' ', COLS - 2);
-		mvwaddnwstr(rover.window, i + 1, 2, WBUF, COLS - 4);
+		mvwaddnwstr(rover.window, i + 1, 2, wbuffer, COLS - 4);
 		if (marking && MARKED(j)) {
 			wcolor_set(rover.window, RVC_MARKS, NULL);
 			mvwaddch(rover.window, i + 1, 1, RVS_MARK);
@@ -374,36 +323,19 @@ void update_view()
 		wcolor_set(rover.window, RVC_SCROLLBAR, NULL);
 		mvwvline(rover.window, center - height / 2 + 1, COLS - 1, RVS_SCROLLBAR, height);
 	}
-	BUF1[0] = FLAGS & SHOW_FILES ? 'F' : ' ';
-	BUF1[1] = FLAGS & SHOW_DIRS ? 'D' : ' ';
-	BUF1[2] = FLAGS & SHOW_HIDDEN ? 'H' : ' ';
+	buffer_one[0] = FLAGS & SHOW_FILES ? 'F' : ' ';
+	buffer_one[1] = FLAGS & SHOW_DIRS ? 'D' : ' ';
+	buffer_one[2] = FLAGS & SHOW_HIDDEN ? 'H' : ' ';
 	if (!rover.nfiles)
-		strcpy(BUF2, "0/0");
+		strcpy(buffer_two, "0/0");
 	else
-		snprintf(BUF2, BUFLEN, "%d/%d", ESEL + 1, rover.nfiles);
-	snprintf(BUF1 + 3, BUFLEN, "%12s", BUF2);
+		snprintf(buffer_two, PATH_MAX, "%d/%d", ESEL + 1, rover.nfiles);
+	snprintf(buffer_one + 3, PATH_MAX, "%12s", buffer_two);
 	color_set(RVC_STATUS, NULL);
-	mvaddstr(LINES - 1, STATUSPOS, BUF1);
+	mvaddstr(LINES - 1, STATUSPOS, buffer_one);
 	wrefresh(rover.window);
 }
 
-/* Show a message on the status bar. */
-void message(Color color, char *fmt, ...)
-{
-	int len, pos;
-	va_list args;
-
-	va_start(args, fmt);
-	vsnprintf(BUF1, MIN(BUFLEN, STATUSPOS), fmt, args);
-	va_end(args);
-	len = strlen(BUF1);
-	pos = (STATUSPOS - len) / 2;
-	attr_on(A_BOLD, NULL);
-	color_set(color, NULL);
-	mvaddstr(LINES - 1, pos, BUF1);
-	color_set(DEFAULT, NULL);
-	attr_off(A_BOLD, NULL);
-}
 
 /* Clear message area, leaving only status info. */
 void clear_message()
@@ -535,10 +467,12 @@ void try_to_sel(const char *target)
 /* Reload CWD, but try to keep selection. */
 void reload()
 {
+	char input[PATH_MAX];
+
 	if (rover.nfiles) {
-		strcpy(INPUT, ENAME(ESEL));
+		strcpy(input, ENAME(ESEL));
 		cd(0);
-		try_to_sel(INPUT);
+		try_to_sel(input);
 		update_view();
 	} else
 		cd(1);
@@ -640,8 +574,7 @@ int process_dir(PROCESS pre, PROCESS proc, PROCESS pos, const char *path)
 /* Process all marked entries using CWD as destination root.
    All marked entries that are directories will be recursively processed.
    See process_dir() for details on the parameters. */
-void process_marked(PROCESS pre, PROCESS proc, PROCESS pos,
-               const char *msg_doing, const char *msg_done)
+void process_marked(PROCESS pre, PROCESS proc, PROCESS pos, const char *msg_doing, const char *msg_done)
 {
 	int i, ret;
 	char *entry;
@@ -703,8 +636,6 @@ int delfile(const char *path)
 	return unlink(path);
 }
 
-PROCESS deldir = rmdir;
-
 int addfile(const char *path)
 {
 	/* Using creat(2) because mknod(2) doesn't seem to be portable. */
@@ -721,8 +652,7 @@ int cpyfile(const char *srcpath)
 	int src, dst, ret;
 	size_t size;
 	struct stat st;
-	char buf[BUFSIZ];
-	char dstpath[PATH_MAX];
+	char buf[BUFSIZ], dstpath[PATH_MAX], buffer[PATH_MAX];
 
 	strcpy(dstpath, CWD);
 	strcat(dstpath, srcpath + strlen(rover.marks.dirpath));
@@ -730,11 +660,11 @@ int cpyfile(const char *srcpath)
 	if (ret < 0)
 		return ret;
 	if (S_ISLNK(st.st_mode)) {
-		ret = readlink(srcpath, BUF1, BUFLEN - 1);
+		ret = readlink(srcpath, buffer, PATH_MAX - 1);
 		if (ret < 0)
 			return ret;
-		BUF1[ret] = '\0';
-		ret       = symlink(BUF1, dstpath);
+		buffer[ret] = '\0';
+		ret       = symlink(buffer, dstpath);
 	} else {
 		ret = src = open(srcpath, O_RDONLY);
 		if (ret < 0)
@@ -790,11 +720,13 @@ int movfile(const char *srcpath)
 
 void start_line_edit(const char *init_input)
 {
+	char input[PATH_MAX];
+
 	curs_set(TRUE);
-	strncpy(INPUT, init_input, BUFLEN);
-	rover.edit.left           = mbstowcs(rover.edit.buffer, init_input, BUFLEN);
-	rover.edit.right          = BUFLEN - 1;
-	rover.edit.buffer[BUFLEN] = L'\0';
+	strncpy(input, init_input, PATH_MAX);
+	rover.edit.left           = mbstowcs(rover.edit.buffer, init_input, PATH_MAX);
+	rover.edit.right          = PATH_MAX - 1;
+	rover.edit.buffer[PATH_MAX] = L'\0';
 	rover.edit_scroll         = 0;
 }
 
@@ -803,6 +735,7 @@ EditStat get_line_edit()
 {
 	wchar_t eraser, killer, wch;
 	int ret, length;
+	char input[PATH_MAX];
 
 	ret = rover_get_wch((wint_t *)&wch);
 	erasewchar(&eraser);
@@ -848,529 +781,11 @@ EditStat get_line_edit()
 				EDIT_INSERT(rover.edit, wch);
 		}
 	}
-	/* Encode edit contents in INPUT. */
+	/* Encode edit contents in input. */
 	rover.edit.buffer[rover.edit.left] = L'\0';
-	length                             = wcstombs(INPUT, rover.edit.buffer, BUFLEN);
-	wcstombs(&INPUT[length], &rover.edit.buffer[rover.edit.right + 1],
-	         BUFLEN - length);
+	length                             = wcstombs(input, rover.edit.buffer, PATH_MAX);
+	wcstombs(&input[length], &rover.edit.buffer[rover.edit.right + 1],
+	         PATH_MAX - length);
 	return CONTINUE;
 }
 
-/* Update line input on the screen. */
-void update_input(const char *prompt, Color color)
-{
-	int plen, ilen, maxlen;
-
-	plen   = strlen(prompt);
-	ilen   = mbstowcs(NULL, INPUT, 0);
-	maxlen = STATUSPOS - plen - 2;
-	if (ilen - rover.edit_scroll < maxlen)
-		rover.edit_scroll = MAX(ilen - maxlen, 0);
-	else if (rover.edit.left > rover.edit_scroll + maxlen - 1)
-		rover.edit_scroll = rover.edit.left - maxlen;
-	else if (rover.edit.left < rover.edit_scroll)
-		rover.edit_scroll = MAX(rover.edit.left - maxlen, 0);
-	color_set(RVC_PROMPT, NULL);
-	mvaddstr(LINES - 1, 0, prompt);
-	color_set(color, NULL);
-	mbstowcs(WBUF, INPUT, COLS);
-	mvaddnwstr(LINES - 1, plen, &WBUF[rover.edit_scroll], maxlen);
-	mvaddch(LINES - 1, plen + MIN(ilen - rover.edit_scroll, maxlen + 1), ' ');
-	color_set(DEFAULT, NULL);
-	if (rover.edit_scroll)
-		mvaddch(LINES - 1, plen - 1, '<');
-	if (ilen > rover.edit_scroll + maxlen)
-		mvaddch(LINES - 1, plen + maxlen, '>');
-	move(LINES - 1, plen + rover.edit.left - rover.edit_scroll);
-}
-
-int main(int argc, char *argv[])
-{
-	int i, ch;
-	char *program;
-	char *entry;
-	const char *key;
-	const char *clip_path;
-	DIR *d;
-	EditStat edit_stat;
-	FILE *save_cwd_file   = NULL;
-	FILE *save_marks_file = NULL;
-	FILE *clip_file;
-
-	if (argc >= 2) {
-		if (!strcmp(argv[1], "-v") || !strcmp(argv[1], "--version")) {
-			printf("rover %s\n", RV_VERSION);
-			return 0;
-		} else if (!strcmp(argv[1], "-h") || !strcmp(argv[1], "--help")) {
-			printf(
-				"Usage: rover [OPTIONS] [DIR [DIR [...]]]\n"
-				"       Browse current directory or the ones specified.\n\n"
-				"  or:  rover -h|--help\n"
-				"       Print this help message and exit.\n\n"
-				"  or:  rover -v|--version\n"
-				"       Print program version and exit.\n\n"
-				"See rover(1) for more information.\n"
-				"Rover homepage: <https://github.com/lecram/rover>.\n");
-			return 0;
-		} else if (!strcmp(argv[1], "-d") || !strcmp(argv[1], "--save-cwd")) {
-			if (argc > 2) {
-				save_cwd_file = fopen(argv[2], "w");
-				argc -= 2;
-				argv += 2;
-			} else {
-				fprintf(stderr, "error: missing argument to %s\n", argv[1]);
-				return 1;
-			}
-		} else if (!strcmp(argv[1], "-m") || !strcmp(argv[1], "--save-marks")) {
-			if (argc > 2) {
-				save_marks_file = fopen(argv[2], "a");
-				argc -= 2;
-				argv += 2;
-			} else {
-				fprintf(stderr, "error: missing argument to %s\n", argv[1]);
-				return 1;
-			}
-		}
-	}
-	get_user_programs();
-	init_term();
-	rover.nfiles = 0;
-	for (i = 0; i < 10; i++) {
-		rover.tabs[i].esel = rover.tabs[i].scroll = 0;
-		rover.tabs[i].flags                       = RV_FLAGS;
-	}
-	strcpy(rover.tabs[0].cwd, getenv("HOME"));
-	for (i = 1; i < argc && i < 10; i++) {
-		if ((d = opendir(argv[i]))) {
-			realpath(argv[i], rover.tabs[i].cwd);
-			closedir(d);
-		} else
-			strcpy(rover.tabs[i].cwd, rover.tabs[0].cwd);
-	}
-	getcwd(rover.tabs[i].cwd, PATH_MAX);
-	for (i++; i < 10; i++)
-		strcpy(rover.tabs[i].cwd, rover.tabs[i - 1].cwd);
-	for (i = 0; i < 10; i++)
-		if (rover.tabs[i].cwd[strlen(rover.tabs[i].cwd) - 1] != '/')
-			strcat(rover.tabs[i].cwd, "/");
-	rover.tab    = 1;
-	rover.window = subwin(stdscr, LINES - 2, COLS, 1, 0);
-	init_marks(&rover.marks);
-	cd(1);
-	strcpy(CLIPBOARD, CWD);
-	if (rover.nfiles > 0)
-		strcat(CLIPBOARD, ENAME(ESEL));
-	while (1) {
-		ch  = rover_getch();
-		key = keyname(ch);
-		clear_message();
-		if (!strcmp(key, RVK_QUIT))
-			break;
-		else if (ch >= '0' && ch <= '9') {
-			rover.tab = ch - '0';
-			cd(0);
-		} else if (!strcmp(key, RVK_HELP)) {
-			spawn((char *[]){ "man", "rover", NULL });
-		} else if (!strcmp(key, RVK_DOWN)) {
-			if (!rover.nfiles)
-				continue;
-			ESEL = MIN(ESEL + 1, rover.nfiles - 1);
-			update_view();
-		} else if (!strcmp(key, RVK_UP)) {
-			if (!rover.nfiles)
-				continue;
-			ESEL = MAX(ESEL - 1, 0);
-			update_view();
-		} else if (!strcmp(key, RVK_JUMP_DOWN)) {
-			if (!rover.nfiles)
-				continue;
-			ESEL = MIN(ESEL + RV_JUMP, rover.nfiles - 1);
-			if (rover.nfiles > HEIGHT)
-				SCROLL = MIN(SCROLL + RV_JUMP, rover.nfiles - HEIGHT);
-			update_view();
-		} else if (!strcmp(key, RVK_JUMP_UP)) {
-			if (!rover.nfiles)
-				continue;
-			ESEL   = MAX(ESEL - RV_JUMP, 0);
-			SCROLL = MAX(SCROLL - RV_JUMP, 0);
-			update_view();
-		} else if (!strcmp(key, RVK_JUMP_TOP)) {
-			if (!rover.nfiles)
-				continue;
-			ESEL = 0;
-			update_view();
-		} else if (!strcmp(key, RVK_JUMP_BOTTOM)) {
-			if (!rover.nfiles)
-				continue;
-			ESEL = rover.nfiles - 1;
-			update_view();
-		} else if (!strcmp(key, RVK_CD_DOWN)) {
-			if (!rover.nfiles || !S_ISDIR(EMODE(ESEL)))
-				continue;
-			if (chdir(ENAME(ESEL)) == -1) {
-				message(RED, "Cannot access \"%s\".", ENAME(ESEL));
-				continue;
-			}
-			strcat(CWD, ENAME(ESEL));
-			cd(1);
-		} else if (!strcmp(key, RVK_CD_UP)) {
-			char *dirname, first;
-			if (!strcmp(CWD, "/"))
-				continue;
-			CWD[strlen(CWD) - 1] = '\0';
-			dirname              = strrchr(CWD, '/') + 1;
-			first                = dirname[0];
-			dirname[0]           = '\0';
-			cd(1);
-			dirname[0]               = first;
-			dirname[strlen(dirname)] = '/';
-			try_to_sel(dirname);
-			dirname[0] = '\0';
-			if (rover.nfiles > HEIGHT)
-				SCROLL = ESEL - HEIGHT / 2;
-			update_view();
-		} else if (!strcmp(key, RVK_HOME)) {
-			strcpy(CWD, getenv("HOME"));
-			if (CWD[strlen(CWD) - 1] != '/')
-				strcat(CWD, "/");
-			cd(1);
-		} else if (!strcmp(key, RVK_TARGET)) {
-			char *bname, first;
-			int is_dir  = S_ISDIR(EMODE(ESEL));
-			ssize_t len = readlink(ENAME(ESEL), BUF1, BUFLEN - 1);
-			if (len == -1)
-				continue;
-			BUF1[len] = '\0';
-			if (access(BUF1, F_OK) == -1) {
-				char *msg;
-				switch (errno) {
-				case EACCES:
-					msg = "Cannot access \"%s\".";
-					break;
-				case ENOENT:
-					msg = "\"%s\" does not exist.";
-					break;
-				default:
-					msg = "Cannot navigate to \"%s\".";
-				}
-				strcpy(BUF2, BUF1); /* message() uses BUF1. */
-				message(RED, msg, BUF2);
-				continue;
-			}
-			realpath(BUF1, CWD);
-			len = strlen(CWD);
-			if (CWD[len - 1] == '/')
-				CWD[len - 1] = '\0';
-			bname  = strrchr(CWD, '/') + 1;
-			first  = *bname;
-			*bname = '\0';
-			cd(1);
-			*bname = first;
-			if (is_dir)
-				strcat(CWD, "/");
-			try_to_sel(bname);
-			*bname = '\0';
-			update_view();
-		} else if (!strcmp(key, RVK_COPY_PATH)) {
-			clip_path = getenv("CLIP");
-			if (!clip_path)
-				goto copy_path_fail;
-			clip_file = fopen(clip_path, "w");
-			if (!clip_file)
-				goto copy_path_fail;
-			fprintf(clip_file, "%s%s\n", CWD, ENAME(ESEL));
-			fclose(clip_file);
-			goto copy_path_done;
-copy_path_fail:
-			strcpy(CLIPBOARD, CWD);
-			strcat(CLIPBOARD, ENAME(ESEL));
-copy_path_done:;
-		} else if (!strcmp(key, RVK_PASTE_PATH)) {
-			clip_path = getenv("CLIP");
-			if (!clip_path)
-				goto paste_path_fail;
-			clip_file = fopen(clip_path, "r");
-			if (!clip_file)
-				goto paste_path_fail;
-			fscanf(clip_file, "%s\n", CLIPBOARD);
-			fclose(clip_file);
-paste_path_fail:
-			strcpy(BUF1, CLIPBOARD);
-			strcpy(CWD, dirname(BUF1));
-			if (strcmp(CWD, "/"))
-				strcat(CWD, "/");
-			cd(1);
-			strcpy(BUF1, CLIPBOARD);
-			try_to_sel(strstr(CLIPBOARD, basename(BUF1)));
-			update_view();
-		} else if (!strcmp(key, RVK_REFRESH)) {
-			reload();
-		} else if (!strcmp(key, RVK_SHELL)) {
-			program = user_shell;
-			if (program) {
-#ifdef RV_SHELL
-				spawn((char *[]){ RV_SHELL, "-c", program, NULL });
-#else
-				spawn((char *[]){ program, NULL });
-#endif
-				reload();
-			}
-		} else if (!strcmp(key, RVK_VIEW)) {
-			if (!rover.nfiles || S_ISDIR(EMODE(ESEL)))
-				continue;
-			if (open_with_env(user_pager, ENAME(ESEL)))
-				cd(0);
-		} else if (!strcmp(key, RVK_EDIT)) {
-			if (!rover.nfiles || S_ISDIR(EMODE(ESEL)))
-				continue;
-			if (open_with_env(user_editor, ENAME(ESEL)))
-				cd(0);
-		} else if (!strcmp(key, RVK_OPEN)) {
-			if (!rover.nfiles || S_ISDIR(EMODE(ESEL)))
-				continue;
-			if (open_with_env(user_open, ENAME(ESEL)))
-				cd(0);
-		} else if (!strcmp(key, RVK_SEARCH)) {
-			int oldsel, oldscroll, length;
-			if (!rover.nfiles)
-				continue;
-			oldsel    = ESEL;
-			oldscroll = SCROLL;
-			start_line_edit("");
-			update_input(RVP_SEARCH, RED);
-			while ((edit_stat = get_line_edit()) == CONTINUE) {
-				int sel;
-				Color color = RED;
-				length      = strlen(INPUT);
-				if (length) {
-					for (sel = 0; sel < rover.nfiles; sel++)
-						if (!strncmp(ENAME(sel), INPUT, length))
-							break;
-					if (sel < rover.nfiles) {
-						color = GREEN;
-						ESEL  = sel;
-						if (rover.nfiles > HEIGHT) {
-							if (sel < 3)
-								SCROLL = 0;
-							else if (sel - 3 > rover.nfiles - HEIGHT)
-								SCROLL = rover.nfiles - HEIGHT;
-							else
-								SCROLL = sel - 3;
-						}
-					}
-				} else {
-					ESEL   = oldsel;
-					SCROLL = oldscroll;
-				}
-				update_view();
-				update_input(RVP_SEARCH, color);
-			}
-			if (edit_stat == CANCEL) {
-				ESEL   = oldsel;
-				SCROLL = oldscroll;
-			}
-			clear_message();
-			update_view();
-		} else if (!strcmp(key, RVK_TG_FILES)) {
-			FLAGS ^= SHOW_FILES;
-			reload();
-		} else if (!strcmp(key, RVK_TG_DIRS)) {
-			FLAGS ^= SHOW_DIRS;
-			reload();
-		} else if (!strcmp(key, RVK_TG_HIDDEN)) {
-			FLAGS ^= SHOW_HIDDEN;
-			reload();
-		} else if (!strcmp(key, RVK_NEW_FILE)) {
-			int ok = 0;
-			start_line_edit("");
-			update_input(RVP_NEW_FILE, RED);
-			while ((edit_stat = get_line_edit()) == CONTINUE) {
-				int length = strlen(INPUT);
-				ok         = length;
-				for (i = 0; i < rover.nfiles; i++) {
-					if (
-						!strncmp(ENAME(i), INPUT, length) &&
-						(!strcmp(ENAME(i) + length, "") ||
-					     !strcmp(ENAME(i) + length, "/"))) {
-						ok = 0;
-						break;
-					}
-				}
-				update_input(RVP_NEW_FILE, ok ? GREEN : RED);
-			}
-			clear_message();
-			if (edit_stat == CONFIRM) {
-				if (ok) {
-					if (addfile(INPUT) == 0) {
-						cd(1);
-						try_to_sel(INPUT);
-						update_view();
-					} else
-						message(RED, "Could not create \"%s\".", INPUT);
-				} else
-					message(RED, "\"%s\" already exists.", INPUT);
-			}
-		} else if (!strcmp(key, RVK_NEW_DIR)) {
-			int ok = 0;
-			start_line_edit("");
-			update_input(RVP_NEW_DIR, RED);
-			while ((edit_stat = get_line_edit()) == CONTINUE) {
-				int length = strlen(INPUT);
-				ok         = length;
-				for (i = 0; i < rover.nfiles; i++) {
-					if (
-						!strncmp(ENAME(i), INPUT, length) &&
-						(!strcmp(ENAME(i) + length, "") ||
-					     !strcmp(ENAME(i) + length, "/"))) {
-						ok = 0;
-						break;
-					}
-				}
-				update_input(RVP_NEW_DIR, ok ? GREEN : RED);
-			}
-			clear_message();
-			if (edit_stat == CONFIRM) {
-				if (ok) {
-					if (adddir(INPUT) == 0) {
-						cd(1);
-						strcat(INPUT, "/");
-						try_to_sel(INPUT);
-						update_view();
-					} else
-						message(RED, "Could not create \"%s/\".", INPUT);
-				} else
-					message(RED, "\"%s\" already exists.", INPUT);
-			}
-		} else if (!strcmp(key, RVK_RENAME)) {
-			int ok = 0;
-			char *last;
-			int isdir;
-			strcpy(INPUT, ENAME(ESEL));
-			last = INPUT + strlen(INPUT) - 1;
-			if ((isdir = *last == '/'))
-				*last = '\0';
-			start_line_edit(INPUT);
-			update_input(RVP_RENAME, RED);
-			while ((edit_stat = get_line_edit()) == CONTINUE) {
-				int length = strlen(INPUT);
-				ok         = length;
-				for (i = 0; i < rover.nfiles; i++)
-					if (
-						!strncmp(ENAME(i), INPUT, length) &&
-						(!strcmp(ENAME(i) + length, "") ||
-					     !strcmp(ENAME(i) + length, "/"))) {
-						ok = 0;
-						break;
-					}
-				update_input(RVP_RENAME, ok ? GREEN : RED);
-			}
-			clear_message();
-			if (edit_stat == CONFIRM) {
-				if (isdir)
-					strcat(INPUT, "/");
-				if (ok) {
-					if (!rename(ENAME(ESEL), INPUT) && MARKED(ESEL)) {
-						del_mark(&rover.marks, ENAME(ESEL));
-						add_mark(&rover.marks, CWD, INPUT);
-					}
-					cd(1);
-					try_to_sel(INPUT);
-					update_view();
-				} else
-					message(RED, "\"%s\" already exists.", INPUT);
-			}
-		} else if (!strcmp(key, RVK_TG_EXEC)) {
-			if (!rover.nfiles || S_ISDIR(EMODE(ESEL)))
-				continue;
-			if (S_IXUSR & EMODE(ESEL))
-				EMODE(ESEL) &= ~(S_IXUSR | S_IXGRP | S_IXOTH);
-			else
-				EMODE(ESEL) |= S_IXUSR | S_IXGRP | S_IXOTH;
-			if (chmod(ENAME(ESEL), EMODE(ESEL))) {
-				message(RED, "Failed to change mode of \"%s\".", ENAME(ESEL));
-			} else {
-				message(GREEN, "Changed mode of \"%s\".", ENAME(ESEL));
-				update_view();
-			}
-		} else if (!strcmp(key, RVK_DELETE)) {
-			if (rover.nfiles) {
-				message(YELLOW, "Delete \"%s\"? (Y/n)", ENAME(ESEL));
-				if (rover_getch() == 'Y') {
-					const char *name = ENAME(ESEL);
-					int ret          = ISDIR(ENAME(ESEL)) ? deldir(name) : delfile(name);
-					reload();
-					if (ret)
-						message(RED, "Could not delete \"%s\".", ENAME(ESEL));
-				} else
-					clear_message();
-			} else
-				message(RED, "No entry selected for deletion.");
-		} else if (!strcmp(key, RVK_TG_MARK)) {
-			if (MARKED(ESEL))
-				del_mark(&rover.marks, ENAME(ESEL));
-			else
-				add_mark(&rover.marks, CWD, ENAME(ESEL));
-			MARKED(ESEL) = !MARKED(ESEL);
-			ESEL         = (ESEL + 1) % rover.nfiles;
-			update_view();
-		} else if (!strcmp(key, RVK_INVMARK)) {
-			for (i = 0; i < rover.nfiles; i++) {
-				if (MARKED(i))
-					del_mark(&rover.marks, ENAME(i));
-				else
-					add_mark(&rover.marks, CWD, ENAME(i));
-				MARKED(i) = !MARKED(i);
-			}
-			update_view();
-		} else if (!strcmp(key, RVK_MARKALL)) {
-			for (i = 0; i < rover.nfiles; i++)
-				if (!MARKED(i)) {
-					add_mark(&rover.marks, CWD, ENAME(i));
-					MARKED(i) = 1;
-				}
-			update_view();
-		} else if (!strcmp(key, RVK_MARK_DELETE)) {
-			if (rover.marks.nentries) {
-				message(YELLOW, "Delete all marked entries? (Y/n)");
-				if (rover_getch() == 'Y')
-					process_marked(NULL, delfile, deldir, "Deleting", "Deleted");
-				else
-					clear_message();
-			} else
-				message(RED, "No entries marked for deletion.");
-		} else if (!strcmp(key, RVK_MARK_COPY)) {
-			if (rover.marks.nentries) {
-				if (strcmp(CWD, rover.marks.dirpath))
-					process_marked(adddir, cpyfile, NULL, "Copying", "Copied");
-				else
-					message(RED, "Cannot copy to the same path.");
-			} else
-				message(RED, "No entries marked for copying.");
-		} else if (!strcmp(key, RVK_MARK_MOVE)) {
-			if (rover.marks.nentries) {
-				if (strcmp(CWD, rover.marks.dirpath))
-					process_marked(adddir, movfile, deldir, "Moving", "Moved");
-				else
-					message(RED, "Cannot move to the same path.");
-			} else
-				message(RED, "No entries marked for moving.");
-		}
-	}
-	if (rover.nfiles)
-		free_rows(&rover.rows, rover.nfiles);
-	delwin(rover.window);
-	if (save_cwd_file != NULL) {
-		fputs(CWD, save_cwd_file);
-		fclose(save_cwd_file);
-	}
-	if (save_marks_file != NULL) {
-		for (i = 0; i < rover.marks.bulk; i++) {
-			entry = rover.marks.entries[i];
-			if (entry)
-				fprintf(save_marks_file, "%s%s\n", rover.marks.dirpath, entry);
-		}
-		fclose(save_marks_file);
-	}
-	free_marks(&rover.marks);
-	return 0;
-}
